refactor: rename internal/anthropic engine package to internal/afk
afk = Agent Framework Kit / Away From Keyboard. The package name no longer
clashes with the anthropic SDK package, so the tbanthropic import alias is
dropped in favor of plain afk. Pure rename: no behavior change.
